feat(policy): Complete Phase 1 SSG conversions! 🎉
PHASE 1 - CONVERSION 4/4 COMPLETE ✅

Final conversion:
- Extract accordion state & language logic to PolicyPageClient.tsx
- Create server component with proper policy & documents SEO
- All functionality preserved: accordion toggles, language context, translations
- Dynamic document links and WhatsApp integration maintained

Results:
✅ Build successful - no client component errors
✅ Static HTML generated (.next/server/app/about/policy.html - 17.4kB)
✅ SEO meta: 'Access Teddy Kids policies, GGD inspection reports...'
✅ Bundle performance stable

🎊 PHASE 1 COMPLETE - ALL 4 PAGES CONVERTED TO SSG!

📊 Final Results:
✅ /about: Server-rendered with proper SEO
✅ /apie: Complex playground now static + SEO
✅ /accepted: Most complex page (confetti/params) → SSG
✅ /about/policy: Accordion state + i18n → SSG

Next: Phase 2 Infrastructure Enhancements 🚀
